feat-#72: refactor activity module
