+++ src/microservices/catalog/graphql/input/create-activity.input.ts
