+++ src/microservices/catalog/graphql/input/update-activity.input.ts
