+++ src/microservices/catalog/graphql/models/gql-activity.model.ts
