+++ src/microservices/catalog/graphql/models/index.ts
@@ -1,0 +1,2 @@
+export * from './gql-activity.model';
+export * from './gql-text-response.model';
