+++ src/microservices/catalog/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './update-activity.interface';
