+++ src/microservices/catalog/rest/models/index.ts
@@ -1,0 +1,2 @@
+export * from './rest-activity.model';
+export * from './rest-text-response.model';
