+++ src/microservices/catalog/rest/models/rest-activity.model.ts
