+++ src/microservices/catalog/services/index.ts
@@ -1,5 +1,1 @@
-import { ActivityService } from './activity/activity.service';
-
-export const CatalogServices = [
-  ActivityService,
-];
+export * from './activity.service';
